+++ microbench_1N/launch_local_multicast.sh
@@ -23,6 +23,13 @@ else
    exit 0
 fi
 
+OUTPUT_DIR="microbench_local_multicast_${NB_CONSUMERS}consumers_${DURATION_XP}sec_${MSG_SIZE}B_${SIZE_BUFFER_LM}messages_in_buffer"
+
+if [ -d $OUTPUT_DIR ]; then
+   echo Local Multicast ${NB_CONSUMERS} consumers, ${DURATION_XP} sec, ${MSG_SIZE}B already done
+   exit 0
+fi
+
 
 rm -rf $MEMORY_DIR && mkdir $MEMORY_DIR
 
@@ -51,7 +58,6 @@ sudo pkill profiler
 ./stop_all.sh
 
 # save files
-OUTPUT_DIR="microbench_local_multicast_${NB_CONSUMERS}consumers_${DURATION_XP}sec_${MSG_SIZE}B_{SIZE_BUFFER_LM}messages_in_buffer"
 mkdir $OUTPUT_DIR
 mv $MEMORY_DIR $OUTPUT_DIR/
 mv statistics*.log $OUTPUT_DIR/
